+++ CotEditor/Sources/Text Finder/Multiple Replace/Models/NSTextView+MultipleReplace.swift
@@ -90,7 +90,7 @@ extension NSTextView {
     ///   - inSelection: Whether find string only in selectedRanges.
     /// - Returns: A result message.
     /// - Throws: `CancellationError`
-    final func replaceAll(_ definition: MultipleReplace, inSelection: Bool) async throws -> String {
+    @discardableResult final func replaceAll(_ definition: MultipleReplace, inSelection: Bool) async throws -> String {
         
         self.isEditable = false
         defer { self.isEditable = true }