+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Allow defining multiple comment delimiters in a syntax definition.
 - Add the Multiline option to syntax highlighting rules.
 - Add a Delimiters pane to the syntax editor for customizing parsing rules.
+- Show your current position in search results in the Find field while navigating between matches (for example, “2/5”).
 
 
 ### Compatibility Changes